Made SpriteLoader script
making it so we can replace just the sprites in the custom game instead of the whole prefab as that proved extremely difficult
